POSHI-491 Share the class
