+++ modules/test/poshi/poshi-runner/src/main/java/com/liferay/poshi/runner/util/LogFileUtil.java
@@ -14,8 +14,6 @@
 
 package com.liferay.poshi.runner.util;
 
-import com.liferay.poshi.core.util.PropsValues;
-import com.liferay.poshi.runner.exception.PoshiRunnerWarningException;
 import com.liferay.poshi.runner.selenium.LiferaySeleniumUtil;
 
 /**
@@ -53,64 +51,6 @@ public class LogFileUtil {
 			text);
 
 		consoleTextPresentCondition.waitFor();
-	}
-
-	public abstract static class Condition {
-
-		public Condition() {
-			this("");
-		}
-
-		public Condition(String message) {
-			_message = message;
-		}
-
-		public void assertTrue() throws Exception {
-			if (!evaluate()) {
-				throw new Exception(_message);
-			}
-		}
-
-		public abstract boolean evaluate() throws Exception;
-
-		public void verify() throws Exception {
-			if (!evaluate()) {
-				throw new PoshiRunnerWarningException(
-					"VERIFICATION_WARNING: " + _message);
-			}
-		}
-
-		public void waitFor() throws Exception {
-			waitFor("true");
-		}
-
-		public void waitFor(String throwException) throws Exception {
-			int timeout = PropsValues.TIMEOUT_EXPLICIT_WAIT * 1000;
-			int wait = 500;
-
-			for (int millisecond = 0; millisecond < timeout;
-				 millisecond += wait) {
-
-				try {
-					if (evaluate()) {
-						return;
-					}
-				}
-				catch (Exception exception) {
-				}
-
-				Thread.sleep(wait);
-			}
-
-			if ((throwException == null) ||
-				Boolean.parseBoolean(throwException)) {
-
-				assertTrue();
-			}
-		}
-
-		private final String _message;
-
 	}
 
 	protected static Condition getConsoleTextNotPresentCondition(String text) {
